% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps.R
\name{lrps}
\alias{lrps}
\title{lrps}
\usage{
lrps(Alpha, Pis, Haz, NP, Ratio = FALSE)
}
\arguments{
\item{NP}{Number of subjects/power of the test.}

\item{alpha}{numeric Two sided type I error.}

\item{haz}{vector hazard rate/ratios of each group.}

\item{pis}{vector probability of each group.}
}
\value{
phsPower/Power/N n
}
\description{
'@description
'This function This function computes the Power or Sample size of a K group test.
'This function is derived from section 3 of Lachin's paper - An ANOVA-like Test
}
\examples{
 lrps(0.05,c(0.335057,0.335057,0.2649026,0.2649026),c(0.0875,0.0875,0.065625,0.065625),5004)
 lrps(0.05,c(0.335057,0.335057,0.2649026,0.2649026),c(0.0875,0.0875,0.065625,0.065625),0.9986035)
 lrps(0.05,c(0.335057,0.335057,0.2649026,0.2649026),c(1,1,0.75,0.75),5000,Ratio=TRUE)
}
