% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps.R
\name{Nmain}
\alias{Nmain}
\title{Sample size/Power}
\usage{
Nmain(alpha, pis, haz, power)
}
\arguments{
\item{alpha}{numeric Two sided type I error.}

\item{pis}{vector probability of each group.}

\item{haz}{vector hazard rate of each group.}

\item{power}{power of the test as desired.}
}
\value{
returns a list of 3 data frames called values, pds, eds. "values" includes a jumbo of statistics includes:
non centrcentrality parameter$chinc=psi^2$, non-centrality factor $ncf=varphi^2$ , Total sample size N, sample size within each group n, total event.
"Pds" includes the expected probabilities under alternative hypothesis. "Eds" calculates # of subjects should have the event"
}
\description{
' @description This main function computes the sample size (N) needed for a k-group log rank test based on parameter inputed.
}
\examples{
Nmain(0.05,c(0.3350577,0.3350577,0.2649026,0.2649026),c(0.0875,0.0875,0.065625,0.065625),0.9986035)
}
