% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{Pi}
\alias{Pi}
\title{Pi}
\usage{
Pi(Rct, T, EHR, THR, G = NULL)
}
\arguments{
\item{Rct}{numeric (R) Years of recruitment}

\item{T}{numeric (T) Total duration of study (includes followup)}

\item{EHR}{numeric Event Hazard rate}

\item{THR}{numeric total Hazard rate(Event hazard rate+loss hazard rate)}

\item{G}{numeric (gamma)The shape parameter for the density function.}

\item{Method}{String, if Method='Unif' then assume the recruitment is uniform entry, else, a gamma value must be provided.}
}
\value{
This function returns a single value that is the expected probability $pi$
}
\description{
' This function calculates $pi$ - the probability that the event is observed under the exponential model.
'This function is derived from section 2 of Lachin's paper: Lachin-Foulkes Exponential Model.
}
\examples{
Pi(1, 7,1.04, 5,-0.27)
Pi(1, 7,1.04, 5)
}
\references{
Lachin JM. Sample size and power for a logrank test and Cox proportional hazards model with multiple groups and strata, or a quantitative covariate with multiple strata. Stat Med. 2013;32(25):4413-4425. doi:10.1002/sim.5839
}
\author{
H. Li, E. Kawaguchi
}
