% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerInt.R
\name{PowerInt}
\alias{PowerInt}
\title{PowerPH}
\usage{
PowerInt(Ratios, Evt, Sigma, Alpha)
}
\arguments{
\item{Ratios}{Numeric - The start point of hazard ratio per standard deviation change.}

\item{Evt}{Numeric - The end point of hazard ratio per standard deviation change.}

\item{Sigma}{Numeric - Numeric - Standard deviation $sigma_k$ of the covariate.}

\item{Alpha}{Numeric - Two sided Type I error $iff P(H_a|H_0)$}
}
\value{
This function returns power of the Cochran's Q test for a K-group test by quantitative covariate interaction.
}
\description{
This function calculates the power of the Cochran's Q test for a K-group test by quantitative covariate interaction.
This function is derived from section 9 of Lachin's paper: Quantitative Covariate Effects.
}
\examples{
PowerInt(c(1.55, 1.45, 1.35, 1.25),394,10,0.05)
}
\references{
Lachin JM. Sample size and power for a logrank test and Cox proportional hazards model with multiple groups and strata, or a quantitative covariate with multiple strata. Stat Med. 2013;32(25):4413-4425. doi:10.1002/sim.5839
}
\author{
H. Li, E. Kawaguchi
}
