% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps.R
\name{Powermain}
\alias{Powermain}
\title{' @description This main function computes Power needed for a k-group log rank test based on parameter inputed.}
\usage{
Powermain(alpha, pis, haz, N)
}
\arguments{
\item{alpha}{numeric Two sided type I error.}

\item{pis}{vector probability of each group.}

\item{haz}{vector hazard rate of each group.}

\item{N}{Number of subjects in the study cohort.}
}
\value{
returns a list of 3 dataframes called values, pds, eds. "values" includes a jumbo of statistics includes:R,T,
non centrcentrality parameter$chinc=psi^2$, non-centrality factor $ncf=varphi^2$ gamma, sample size N, total event.
"Pds" includes the expected probabilities under alternative hypothesis. "Eds calculates # of subjects should have the event"
}
\description{
' @description This main function computes Power needed for a k-group log rank test based on parameter inputed.
}
\examples{
Powermain(0.05,c(1,1,0.75,0.75),0.0875,0.04,3,7,-0.27,5000)
}
