% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerPH.R
\name{PowerPH}
\alias{PowerPH}
\title{PowerPH}
\usage{
PowerPH(Beg, End, Inc, Alpha, Sigma, Evt)
}
\arguments{
\item{Beg}{Numeric - The start point of hazard ratio per standard deviation change.}

\item{End}{Numeric - The end point of hazard ratio per standard deviation change.}

\item{Inc}{Numeric - Unit of increasing of hazard ratio per standard deviation change.}

\item{Alpha}{Numeric - Two sided Type I error iff P(H_a|H_0)$}

\item{Sigma}{Numeric - Standard deviation of the covariate, this is $sigma$ in Lachin's paper.}

\item{Evt}{Numeric - Number of total events expected for N subjects. This can be calculated by $$sum_{j=1}^K N zeta_j pi_j$$ Whereas $zeta_j$ is sample fraction of subjects assigned to each group. $pi_j$ is the event probability calculated by function Prob.}
}
\value{
This function returns list of each increase per unit of hazard ratio and power.
}
\description{
This function calculates the power of association for a quantitative covariate under the Proportional Hazard model.
This function is derived from section 9 of Lachin's paper: Quantitative Covariate Effects.
}
\examples{
PowerPH(1.1,1.4,0.05,0.05,10,1576)
}
\references{
Lachin JM. Sample size and power for a logrank test and Cox proportional hazards model with multiple groups and strata, or a quantitative covariate with multiple strata. Stat Med. 2013;32(25):4413-4425. doi:10.1002/sim.5839
}
\author{
H. Li, E. Kawaguchi
}
