% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps.R
\name{PowerPHmain}
\alias{PowerPHmain}
\title{' @description This main function computes Power needed for a k-group log rank test based on parameter inputed under the proportional hazard model.}
\usage{
PowerPHmain(alpha, pis, rrs, N)
}
\arguments{
\item{alpha}{numeric Two sided type I error.}

\item{pis}{vector probability of each group.}

\item{rrs}{vector hazard ratio of each group.}

\item{N}{Number of subjects in the study cohort.}
}
\value{
returns a list of 3 dataframes called values, pds, eds. "values" includes a PhsPower and total event.
"pis" includes the expected probabilities under alternative hypothesis. "Eds calculates # of subjects should have the event"
}
\description{
' @description This main function computes Power needed for a k-group log rank test based on parameter inputed under the proportional hazard model.
}
\examples{
PowerPHmain(0.05,c(0.335057,0.335057,0.2649026,0.2649026),c(1,1,0.75,0.75),2315)
}
